Move constants to constants/
