+++ src/reducers/__tests__/navigation-test.js
@@ -1,5 +1,5 @@
 import { pushRoute, popRoute } from 'src/navigation/actions';
-import * as constants from 'src/navigation/constants';
+import * as constants from 'src/constants/navigation';
 import reduce from 'src/reducers/navigation';
